внес правки в соотвествии с замечаниями
