+++ js/add-back-button-click.js
@@ -1,0 +1,8 @@
+export default (element, callback) => {
+  const backButton = element.querySelector(`.back`);
+
+  backButton.addEventListener(`click`, () => {
+    callback();
+  });
+};
+
