+++ js/app.js
@@ -19,7 +19,7 @@ class App {
     this.showIntro();
     gameModel.getQuestions()
       .then((games) => this.setup(games))
-      .then(() => this.changeController(this._parseHashFromUrl()), true)
+      .then(() => this.changeController(this._parseHashFromUrl(), true))
       .catch(window.console.error);
   }
 
@@ -31,7 +31,6 @@ class App {
       [ControllerId.STATS]: new Stats()
     };
     window.onhashchange = () => this.changeController(this._parseHashFromUrl());
-
   }
 
   showIntro() {
@@ -48,13 +47,13 @@ class App {
   }
 
   _encodeData(data) {
-    const encode64 = atob(data);
-    return encode64 ? JSON.parse(encode64) : null;
+    const encode64 = encodeURI(data);
+    return encode64 ? encode64 : null;
 
   }
 
   _decodeData(data) {
-    return btoa(JSON.stringify(data));
+    return decodeURI(data);
   }
 
   changeController({controller, value}, isCrossfade) {
