+++ js/data/points.js
@@ -1,4 +1,4 @@
-import {Points, Result} from '../data';
+import {Points, Result} from './data';
 
 export const getPointsByAnswerType = (type, stats) => {
   const countStatType = stats.filter((stat) => stat === type).length;
@@ -10,12 +10,12 @@ export const getPointByLives = (lives) => {
 };
 
 export const getRightPoints = (stats) => {
-  return stats.filter((stat) => stat !== Result.wrong).length * Points[Result.correct];
+  return stats.filter((stat) => stat !== Result.WRONG).length * Points[Result.CORRECT];
 };
 
 export const getTotalPoints = ({lives, stats}) => {
-  const sumFastAswerPoints = getPointsByAnswerType(Result.fast, stats);
-  const sumSlowAswerPoints = getPointsByAnswerType(Result.slow, stats);
+  const sumFastAswerPoints = getPointsByAnswerType(Result.FAST, stats);
+  const sumSlowAswerPoints = getPointsByAnswerType(Result.SLOW, stats);
   const sumLivesPoints = getPointByLives(lives);
   const sumRightPoints = getRightPoints(stats);
   return sumFastAswerPoints + sumSlowAswerPoints + sumLivesPoints + sumRightPoints;
