+++ js/game/game-adapter.js
@@ -1,42 +1,7 @@
-import {DefaultAdapter} from '../model';
-import {QuestionType, AnswerType} from '../data';
+import {QuestionType, AnswerType} from '../data/data';
 import resizeImage from '../data/resizeImage';
-const getAnswersCountByType = (answers, type) => {
-  return answers.filter((answer) => answer.type === type).length;
-};
 
-const getRightAnswerType = (answers) => {
-  return Object.values(AnswerType)
-    .find((type) => getAnswersCountByType(answers, type) === 1);
-};
-
-const cleanWrongAnswerTypes = (answers) => {
-  const type = getRightAnswerType(answers);
-  answers.forEach((answer) => {
-    answer.type = answer.type === type ? answer.type : null;
-    return answer;
-  });
-};
-
-const setImageSize = (img, answerImage, resolve) => {
-  const correctedSizes = resizeImage({
-    width: answerImage.width,
-    height: answerImage.height
-  }, img);
-  answerImage.width = correctedSizes.width;
-  answerImage.height = correctedSizes.height;
-  resolve();
-};
-
-const loadImage = (answerImage) => {
-  return new Promise((resolve) => {
-    const img = new Image();
-    img.addEventListener(`load`, () => setImageSize(img, answerImage, resolve));
-    img.src = answerImage.url;
-  });
-};
-
-export default new class extends DefaultAdapter {
+export default new class {
 
   preprocessQuestions(data) {
     for (const game of data) {
@@ -71,4 +36,41 @@ export default new class extends DefaultAdapter {
     return response.json();
   }
 }();
+
+const getAnswersCountByType = (answers, type) => {
+  return answers.filter((answer) => answer.type === type).length;
+};
+
+const getRightAnswerType = (answers) => {
+  return Object.values(AnswerType)
+    .find((type) => getAnswersCountByType(answers, type) === 1);
+};
+
+const cleanWrongAnswerTypes = (answers) => {
+  const type = getRightAnswerType(answers);
+  answers.forEach((answer) => {
+    answer.type = answer.type === type ? answer.type : null;
+    return answer;
+  });
+};
+
+const setImageSize = (img, answerImage, resolve) => {
+  const correctedSizes = resizeImage({
+    width: answerImage.width,
+    height: answerImage.height
+  }, img);
+  answerImage.width = correctedSizes.width;
+  answerImage.height = correctedSizes.height;
+  resolve();
+};
+
+const loadImage = (answerImage) => {
+  return new Promise((resolve) => {
+    const img = new Image();
+    img.addEventListener(`load`, () => setImageSize(img, answerImage, resolve));
+    img.addEventListener(`error`, () => resolve());
+    img.src = answerImage.url;
+  });
+};
+
