--- js/model.js
@@ -1,11 +1,0 @@
-export class DefaultAdapter {
-  constructor() {
-    if (new.target === DefaultAdapter) {
-      throw new Error();
-    }
-  }
-
-  preprocess(data) {
-    return data;
-  }
-}
