+++ sass/intro.scss
@@ -3,10 +3,10 @@
 .slide {
   opacity: 0;
   position: absolute;
-  -webkit-transition: opacity 4s linear;
-  -moz-transition: opacity 4s linear;
-  -o-transition: opacity 4s linear;
-  transition: opacity 4s linear;
+  -webkit-transition: opacity 2s linear;
+  -moz-transition: opacity 2s linear;
+  -o-transition: opacity 2s linear;
+  transition: opacity 2s linear;
 }
 
 .intro {
